pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "malone-ai"
version = "0.1.0"
description = "A personal AI assistant with voice interaction"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.5.0",
    "numpy>=1.24",
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
    "edge-tts>=6.1.0",
    "miniaudio>=1.59",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]
browser = [
    "playwright>=1.40",
]
network = [
    "paramiko>=3.0",
    "kubernetes>=28.0",
]

[project.scripts]
malone = "malone.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
